feat: refactor chart settings components and introduce ChartSettingsContent for improved structure
