--- packages/mosaic/src/chart/chart-settings/index.ts
@@ -1,9 +1,0 @@
-export {ChartSettings} from './ChartSettings';
-export {
-  ChartSettingsProvider,
-  useChartSettingsContext,
-} from './ChartSettingsContext';
-export {ChartSettingsPanel} from './ChartSettingsPanel';
-export {ChartSpecViewerPanel} from './ChartSpecViewerPanel';
-export {ChartTypeSelector} from './ChartTypeSelector';
-export {useTableColumns} from './useTableColumns';
