+++ packages/mosaic/src/chart/useGenerateSpec.ts
@@ -11,7 +11,6 @@ export type UseGenerateSpecResult =
     }
   | {
       spec?: undefined;
-
       error: ChartSpecError;
     };
